src/frontend/requirements.txt
flask==3.1.0
requests==2.32.3
gunicorn==23.0.0
python-dotenv==1.0.1
azure-monitor-opentelemetry==1.6.5